+++ CHANGES.rst
@@ -2,6 +2,8 @@
 ------------------
 - Added rudimentary debugging to nrdp plugins (#27) -BH
 - Added ability to receive JSON check results (#23) -Box293 (Troy)
+- Added documentation for --usestdin and --delim (#17) -BH
+- Added ability to specify a delimiter when using stdin (#17) -BH
 
 1.4.0 - 01/06/2017
 ------------------